+++ src/calendar/calendar.jsx
@@ -4,7 +4,7 @@
 
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 
-import { autobind } from 'core-decorators';
+import autobind from 'core-decorators/lib/autobind';
 import React from 'react';
 import Type from 'prop-types';
 
@@ -643,7 +643,7 @@ class Calendar extends React.Component {
         }
 
         if (isInitializing || this.props.selectedTo !== nextProps.selectedTo) {
-            this.selectedFrom = nextProps.selectedTo ? normalizeDate(nextProps.selectedTo) : null;
+            this.selectedTo = nextProps.selectedTo ? normalizeDate(nextProps.selectedTo) : null;
         }
 
         if (isInitializing || this.props.selectedFrom !== nextProps.selectedFrom) {