+++ src/link/link.jsx
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { autobind } from 'core-decorators';
+import autobind from 'core-decorators/lib/autobind';
 import React from 'react';
 import Type from 'prop-types';
 
@@ -24,6 +24,8 @@ class Link extends React.Component {
         text: Type.node,
         /** href ссылки */
         url: Type.string,
+        /** Указание на загрузку, вместо открытия и указание имени файла  */
+        download: Type.oneOfType([Type.string, Type.bool]),
         /** target ссылки */
         target: Type.oneOf(['_self', '_blank', '_parent', '_top']),
         /** Последовательность перехода между контролами при нажатии на Tab */
@@ -78,6 +80,7 @@ class Link extends React.Component {
 
         let linkProps = {
             ref: (root) => { this.root = root; },
+            download: this.props.download,
             className: cn({
                 disabled: this.props.disabled,
                 checked: this.props.checked,
@@ -103,6 +106,7 @@ class Link extends React.Component {
             linkProps.href = this.props.url;
             linkProps.target = this.props.target;
         }
+
 
         let linkContent = [this.props.children];
         let iconTemplate = (