+++ src/menu/menu.jsx
@@ -4,7 +4,7 @@
 
 /* eslint jsx-a11y/no-static-element-interactions: 0 */
 
-import { autobind } from 'core-decorators';
+import autobind from 'core-decorators/lib/autobind';
 import React from 'react';
 import styleType from 'react-style-proptype';
 import Type from 'prop-types';
@@ -47,11 +47,11 @@ class Menu extends React.Component {
         content: Type.arrayOf(Type.shape({
             /** Тип элемента */
             type: Type.oneOf(['item', 'group']),
-            /** Только для type='item', свойство для компонента [MenuItem](../menu-item/) */
+            /** Только для type='item', свойство для компонента [MenuItem](#!/MenuItem) */
             value: Type.oneOfType([Type.string, Type.number]),
             /** Содержание элемента */
             content: Type.oneOfType([Type.node, Type.array]),
-            /** Только для type='item': свойства для компонента [MenuItem](../menu-item/) */
+            /** Только для type='item': свойства для компонента [MenuItem](#!/MenuItem) */
             props: Type.object
         })),
         /** Список значений выбранных элементов */